// src/app/features/company/components/add-company/models/company-phone-number.interface.ts
export class PhoneNumber {
    phone_number: string;
    lavi_branchId?: string;
    date_updated?: Date;
    capabilities?: string[];
    account_sid?: string;
    url?: string;
    country_code?: string;
    sid?: string;
    date_created?: Date;
    service_sid?: string;
    oldLaviBranchId?: string;
}